Conflicts:
	data/images/recordsheets/Conventional_Infantry_no_tables.svg
	data/images/recordsheets/Conventional_Infantry_tables.svg
	src/megameklab/com/util/UnitPrintManager.java
	src/megameklab/com/util/UnitUtil.java

+++ src/megameklab/com/ui/Vehicle/StatusBar.java
@@ -167,7 +167,7 @@ public class StatusBar extends ITab {
         currentTonnage = testEntity.calculateWeight();
 
         currentTonnage += UnitUtil.getUnallocatedAmmoTonnage(getTank());
-        int currentCost = (int)Math.round(getTank().getCost(false));
+        long currentCost = (long) Math.round(getTank().getCost(false));
 
         tons.setText("Tonnage: " + currentTonnage + "/" + tonnage);
         tons.setToolTipText("Current Tonnage/Max Tonnage");
